requirements.txt
requests
python-dotenv
openpyxl
tqdm
